requirements.txt
ccxt==3.0.74
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
matplotlib==3.7.2 